backend/requirements.txt
argon2-cffi-bindings==21.2.0 ; python_full_version == "3.9.13"
argon2-cffi==21.3.0 ; python_full_version == "3.9.13"
asgiref==3.6.0 ; python_full_version == "3.9.13"
autopep8==1.7.0 ; python_full_version == "3.9.13"
certifi==2023.5.7 ; python_full_version == "3.9.13"
cffi==1.15.1 ; python_full_version == "3.9.13"
charset-normalizer==3.1.0 ; python_full_version == "3.9.13"
coreapi==2.3.3 ; python_full_version == "3.9.13"
coreschema==0.0.4 ; python_full_version == "3.9.13"
cryptography==41.0.1 ; python_full_version == "3.9.13"
django-ckeditor==6.5.1 ; python_full_version == "3.9.13"
django-cors-headers==4.1.0 ; python_full_version == "3.9.13"
django-filebrowser==4.0.2 ; python_full_version == "3.9.13"
django-grappelli==3.0.6 ; python_full_version == "3.9.13"
django-js-asset==2.0.0 ; python_full_version == "3.9.13"
django-rest-knox==4.2.0 ; python_full_version == "3.9.13"
django-tinymce==3.6.1 ; python_full_version == "3.9.13"
django==4.2.1 ; python_full_version == "3.9.13"
djangorestframework==3.14.0 ; python_full_version == "3.9.13"
drf-yasg==1.21.5 ; python_full_version == "3.9.13"
idna==3.4 ; python_full_version == "3.9.13"
inflection==0.5.1 ; python_full_version == "3.9.13"
itypes==1.2.0 ; python_full_version == "3.9.13"
jinja2==3.1.2 ; python_full_version == "3.9.13"
markupsafe==2.1.3 ; python_full_version == "3.9.13"
packaging==23.1 ; python_full_version == "3.9.13"
pillow==9.5.0 ; python_full_version == "3.9.13"
psycopg2-binary==2.9.6 ; python_full_version == "3.9.13"
pycodestyle==2.10.0 ; python_full_version == "3.9.13"
pycparser==2.21 ; python_full_version == "3.9.13"
pydantic==1.10.7 ; python_full_version == "3.9.13"
python-dotenv==0.21.1 ; python_full_version == "3.9.13"
pytz==2023.3 ; python_full_version == "3.9.13"
requests==2.31.0 ; python_full_version == "3.9.13"
ruamel-yaml-clib==0.2.7 ; platform_python_implementation == "CPython" and python_full_version == "3.9.13"
ruamel-yaml==0.17.31 ; python_full_version == "3.9.13"
six==1.16.0 ; python_full_version == "3.9.13"
sqlparse==0.4.4 ; python_full_version == "3.9.13"
toml==0.10.2 ; python_full_version == "3.9.13"
typing-extensions==4.5.0 ; python_full_version == "3.9.13"
tzdata==2023.3 ; sys_platform == "win32" and python_full_version == "3.9.13"
uritemplate==4.1.1 ; python_full_version == "3.9.13"
urllib3==2.0.2 ; python_full_version == "3.9.13"
